refactor: rename implementation files and add index re-exports
- Rename component files to match their directory names.
- Add index.ts files to handle re-exports and simplify import paths.
