+++ src/components/cities-list/index.ts
@@ -1,0 +1,1 @@
+export { default } from './cities-list';
