+++ src/components/header/header.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { withHistory, withStore } from '../../utils/mock-component';
-import Header from './index';
+import Header from './header';
 import { AuthorizationStatus, RequestStatus, NameSpace } from '../../const';
 import { makeFakeUser } from '../../utils/mocks';
 import { fetchFavoritesAction } from '../../store/favorites';
