+++ src/components/main-empty/index.ts
@@ -1,0 +1,1 @@
+export { default } from './main-empty';
