+++ src/components/offer-card/index.ts
@@ -1,0 +1,1 @@
+export { default } from './offer-card';
