+++ src/components/offer-list/index.ts
@@ -1,0 +1,1 @@
+export { default } from './offer-list';
