+++ src/components/offer-list/offer-list.tsx
