+++ src/components/private-route/index.ts
@@ -1,0 +1,1 @@
+export { default } from './private-route';
