+++ src/components/sort-options/index.ts
@@ -1,0 +1,1 @@
+export { default } from './sort-options';
