+++ src/components/spinner/spinner.test.tsx
@@ -1,5 +1,5 @@
 import { render } from '@testing-library/react';
-import Spinner from './index';
+import Spinner from './spinner';
 
 describe('Component: Spinner', () => {
   it('should render correctly', () => {
