+++ src/pages/favorites-page/index.ts
@@ -1,0 +1,1 @@
+export { default } from './favorites-page';
