+++ src/pages/login-page/index.ts
@@ -1,0 +1,1 @@
+export { default } from './login-page';
