+++ src/pages/main-page/index.ts
@@ -1,0 +1,1 @@
+export { default } from './main-page';
