+++ src/pages/not-found-page/index.ts
@@ -1,0 +1,1 @@
+export { default } from './not-found-page';
