+++ src/pages/not-found-page/not-found-page.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import { describe, it, expect } from 'vitest';
-import NotFoundPage from './index';
+import NotFoundPage from './not-found-page';
 import { withHistory } from '../../utils/mock-component';
 
 describe('Page: NotFoundPage', () => {
