+++ src/pages/offer-page/offer-page.tsx
